Reorganized folder structure.
