+++ Bibliografia/Paper_Ramos_et_al_GRL2006.pdf
